pyproject.toml
[tool.poetry]
name = "flask-portfolio"
version = "0.1.0"
description = "Flask portfolio created by Oleh Skupeiko"
authors = ["Oleh Skupeiko <sov30082003@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Flask-Migrate = "^4.0.2"
flask-sqlalchemy = "^3.0.2"
Flask-Login = "^0.6.2"
psycopg2 = "^2.9.5"
SQLAlchemy = "^1.4.46"
python-dotenv = "^0.21.1"
email-validator = "^1.3.1"
Flask-WTF = "^1.1.1"
Werkzeug = "^2.2.2"
click = "^8.1.3"
MarkupSafe = "^2.1.2"
Flask-Mail = "^0.9.1"
pydantic = "^1.10.4"
wtforms_sqlalchemy = "^0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^22.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
